+++ main/src/main/java/com/sedmelluq/discord/lavaplayer/source/youtube/YoutubeAudioSourceManager.java
@@ -47,6 +47,7 @@ public class YoutubeAudioSourceManager implements AudioSourceManager, HttpConfig
   private final boolean allowSearch;
   private final YoutubeTrackDetailsLoader trackDetailsLoader;
   private final YoutubeSearchResultLoader searchResultLoader;
+  private final YoutubeSearchMusicResultLoader searchMusicResultLoader;
   private final YoutubePlaylistLoader playlistLoader;
   private final YoutubeLinkRouter linkRouter;
   private final LoadingRoutes loadingRoutes;
@@ -67,6 +68,7 @@ public class YoutubeAudioSourceManager implements AudioSourceManager, HttpConfig
         allowSearch,
         new DefaultYoutubeTrackDetailsLoader(),
         new YoutubeSearchProvider(),
+        new YoutubeSearchMusicProvider(),
         new YoutubeSignatureCipherManager(),
         new DefaultYoutubePlaylistLoader(),
         new DefaultYoutubeLinkRouter(),
@@ -78,6 +80,7 @@ public class YoutubeAudioSourceManager implements AudioSourceManager, HttpConfig
       boolean allowSearch,
       YoutubeTrackDetailsLoader trackDetailsLoader,
       YoutubeSearchResultLoader searchResultLoader,
+      YoutubeSearchMusicResultLoader searchMusicResultLoader,
       YoutubeSignatureResolver signatureResolver,
       YoutubePlaylistLoader playlistLoader,
       YoutubeLinkRouter linkRouter,
@@ -90,6 +93,7 @@ public class YoutubeAudioSourceManager implements AudioSourceManager, HttpConfig
     this.trackDetailsLoader = trackDetailsLoader;
     this.signatureResolver = signatureResolver;
     this.searchResultLoader = searchResultLoader;
+    this.searchMusicResultLoader = searchMusicResultLoader;
     this.playlistLoader = playlistLoader;
     this.linkRouter = linkRouter;
     this.mixLoader = mixLoader;
@@ -97,7 +101,8 @@ public class YoutubeAudioSourceManager implements AudioSourceManager, HttpConfig
 
     combinedHttpConfiguration = new MultiHttpConfigurable(Arrays.asList(
         httpInterfaceManager,
-        searchResultLoader.getHttpConfiguration()
+        searchResultLoader.getHttpConfiguration(),
+        searchMusicResultLoader.getHttpConfiguration()
     ));
   }
 
@@ -184,6 +189,10 @@ public class YoutubeAudioSourceManager implements AudioSourceManager, HttpConfig
     return searchResultLoader.getHttpConfiguration();
   }
 
+  public ExtendedHttpConfigurable getSearchMusicHttpConfiguration() {
+    return searchMusicResultLoader.getHttpConfiguration();
+  }
+
   private AudioItem loadItemOnce(AudioReference reference) {
     return linkRouter.route(reference.identifier, loadingRoutes);
   }
@@ -250,11 +259,15 @@ public class YoutubeAudioSourceManager implements AudioSourceManager, HttpConfig
 
     @Override
     public AudioItem search(String query, Boolean ytMusic) {
-      if (allowSearch) {
+      if (allowSearch && !ytMusic) {
         return searchResultLoader.loadSearchResult(
             query,
-            YoutubeAudioSourceManager.this::buildTrackFromInfo,
-            ytMusic
+            YoutubeAudioSourceManager.this::buildTrackFromInfo
+        );
+      } else if (allowSearch) {
+        return searchMusicResultLoader.loadSearchMusicResult(
+            query,
+            YoutubeAudioSourceManager.this::buildTrackFromInfo
         );
       } else {
         return null;
