+++ main/src/main/java/com/sedmelluq/discord/lavaplayer/source/youtube/YoutubeSearchMusicResultLoader.java
@@ -6,8 +6,8 @@ import com.sedmelluq.discord.lavaplayer.track.AudioTrack;
 import com.sedmelluq.discord.lavaplayer.track.AudioTrackInfo;
 import java.util.function.Function;
 
-public interface YoutubeSearchResultLoader {
-  AudioItem loadSearchResult(String query, Function<AudioTrackInfo, AudioTrack> trackFactory);
+public interface YoutubeSearchMusicResultLoader {
+  AudioItem loadSearchMusicResult(String query, Function<AudioTrackInfo, AudioTrack> trackFactory);
 
   ExtendedHttpConfigurable getHttpConfiguration();
 }
