+++ main/src/main/java/com/sedmelluq/discord/lavaplayer/source/youtube/YoutubeSearchProvider.java
@@ -14,9 +14,7 @@ import com.sedmelluq.discord.lavaplayer.track.AudioTrackInfo;
 import com.sedmelluq.discord.lavaplayer.track.BasicAudioPlaylist;
 import org.apache.http.client.methods.CloseableHttpResponse;
 import org.apache.http.client.methods.HttpGet;
-import org.apache.http.client.methods.HttpPost;
 import org.apache.http.client.utils.URIBuilder;
-import org.apache.http.entity.StringEntity;
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
@@ -41,11 +39,8 @@ public class YoutubeSearchProvider implements YoutubeSearchResultLoader {
   private static final Logger log = LoggerFactory.getLogger(YoutubeSearchProvider.class);
 
   private static final String WATCH_URL_PREFIX = "https://www.youtube.com/watch?v=";
-  private static final String YT_MUSIC_KEY = "AIzaSyC9XL3ZjWddXya6X74dJoCTL-WEYFDNX30";
-  private static final String YT_MUSIC_PAYLOAD = "{\"context\":{\"client\":{\"clientName\":\"WEB_REMIX\",\"clientVersion\":\"0.1\"}},\"query\":\"%s\",\"params\":\"Eg-KAQwIARAAGAAgACgAMABqCBADEAkQBBAK\"}";
   private final HttpInterfaceManager httpInterfaceManager;
   private final Pattern polymerInitialDataRegex = Pattern.compile("(window\\[\"ytInitialData\"]|var ytInitialData)\\s*=\\s*(.*);");
-  private final Pattern ytMusicDataRegex = Pattern.compile("<body>\\s*(.*)\\s*<\\/body>");
 
   public YoutubeSearchProvider() {
     this.httpInterfaceManager = HttpClientTools.createCookielessThreadLocalManager();
@@ -57,57 +52,32 @@ public class YoutubeSearchProvider implements YoutubeSearchResultLoader {
 
   /**
    * @param query Search query.
-   * @param ytMusic boolean
    * @return Playlist of the first page of results.
    */
   @Override
-  public AudioItem loadSearchResult(String query, Function<AudioTrackInfo, AudioTrack> trackFactory, Boolean ytMusic) {
+  public AudioItem loadSearchResult(String query, Function<AudioTrackInfo, AudioTrack> trackFactory) {
     log.debug("Performing a search with query {}", query);
 
     try (HttpInterface httpInterface = httpInterfaceManager.getInterface()) {
-      URI url;
-      if (ytMusic) {
-        url = new URIBuilder("https://music.youtube.com/youtubei/v1/search")
-                .addParameter("alt", "json")
-                .addParameter("key", YT_MUSIC_KEY).build();
-      } else {
-        url = new URIBuilder("https://www.youtube.com/results")
-                .addParameter("search_query", query)
-                .addParameter("hl", "en")
-                .addParameter("persist_hl", "1").build();
-      }
+      log.info("Searching on YouTube");
+      URI url = new URIBuilder("https://www.youtube.com/results")
+          .addParameter("search_query", query)
+          .addParameter("hl", "en")
+          .addParameter("persist_hl", "1").build();
 
-      if (ytMusic) {
-        HttpPost post = new HttpPost(url);
-        StringEntity payload = new StringEntity(String.format(YT_MUSIC_PAYLOAD, query), "UTF-8");
-        post.setHeader("Referer", "music.youtube.com");
-        post.setEntity(payload);
-        try (CloseableHttpResponse response = httpInterface.execute(post)) {
-          return searchQuery(query, trackFactory, response, true);
-        }
-      } else {
-        try (CloseableHttpResponse response = httpInterface.execute(new HttpGet(url))) {
-          return searchQuery(query, trackFactory, response, false);
-        }
+      try (CloseableHttpResponse response = httpInterface.execute(new HttpGet(url))) {
+        HttpClientTools.assertSuccessWithContent(response, "search response");
+
+        Document document = Jsoup.parse(response.getEntity().getContent(), StandardCharsets.UTF_8.name(), "");
+        return extractSearchResults(document, query, trackFactory);
       }
     } catch (Exception e) {
       throw ExceptionTools.wrapUnfriendlyExceptions(e);
     }
   }
 
-  private AudioItem searchQuery(String query, Function<AudioTrackInfo, AudioTrack> trackFactory,
-                                CloseableHttpResponse response, Boolean ytMusic) throws IOException {
-    int statusCode = response.getStatusLine().getStatusCode();
-    if (!HttpClientTools.isSuccessWithContent(statusCode)) {
-      throw new IOException("Invalid status code for search response: " + statusCode);
-    }
-
-    Document document = Jsoup.parse(response.getEntity().getContent(), StandardCharsets.UTF_8.name(), "");
-    return extractSearchResults(document, query, trackFactory, ytMusic);
-  }
-
   private AudioItem extractSearchResults(Document document, String query,
-                                         Function<AudioTrackInfo, AudioTrack> trackFactory, Boolean ytMusic) {
+                                         Function<AudioTrackInfo, AudioTrack> trackFactory) {
 
     List<AudioTrack> tracks = new ArrayList<>();
     Elements resultsSelection = document.select("#page > #content #results");
@@ -118,13 +88,6 @@ public class YoutubeSearchProvider implements YoutubeSearchResultLoader {
             extractTrackFromResultEntry(tracks, result, trackFactory);
           }
         }
-      }
-    } else if (ytMusic) {
-      log.debug("Attempting to parse results from ytMusic page");
-      try {
-        tracks = extractMusicTracks(document, trackFactory);
-      } catch (IOException e) {
-        throw new RuntimeException(e);
       }
     } else {
       log.debug("Attempting to parse results page as polymer");
@@ -209,80 +172,6 @@ public class YoutubeSearchProvider implements YoutubeSearchResultLoader {
 
     long duration = DataFormatTools.durationTextToMillis(lengthText);
     String videoId = renderer.get("videoId").text();
-
-    AudioTrackInfo info = new AudioTrackInfo(title, author, duration, videoId, false,
-        WATCH_URL_PREFIX + videoId);
-
-    return trackFactory.apply(info);
-  }
-
-  private List<AudioTrack> extractMusicTracks(Document document, Function<AudioTrackInfo, AudioTrack> trackFactory) throws IOException {
-    Matcher matcher = ytMusicDataRegex.matcher(document.outerHtml());
-    if (!matcher.find()) {
-      log.warn("Failed to match ytMusicData JSON object");
-      return Collections.emptyList();
-    }
-
-    JsonBrowser jsonBrowser = JsonBrowser.parse(matcher.group(1));
-    ArrayList<AudioTrack> list = new ArrayList<>();
-    jsonBrowser.get("contents")
-        .get("sectionListRenderer")
-        .get("contents")
-        .index(0)
-        .get("musicShelfRenderer")
-        .get("contents")
-        .values()
-        .forEach(json -> {
-          AudioTrack track = extractMusicData(json, trackFactory);
-          if (track != null) list.add(track);
-        });
-    return list;
-  }
-
-  private AudioTrack extractMusicData(JsonBrowser json, Function<AudioTrackInfo, AudioTrack> trackFactory) {
-    JsonBrowser renderer = json.get("musicResponsiveListItemRenderer");
-
-    if (renderer.isNull()) {
-      // Somehow we don't get track info, ignore
-      return null;
-    }
-
-    String title = renderer.get("flexColumns")
-        .index(0)
-        .get("musicResponsiveListItemFlexColumnRenderer")
-        .get("text")
-        .get("runs")
-        .index(0)
-        .get("text").text();
-    String author = renderer.get("flexColumns")
-        .index(1)
-        .get("musicResponsiveListItemFlexColumnRenderer")
-        .get("text")
-        .get("runs")
-        .index(0)
-        .get("text").text();
-    // We do this because sometimes youtube music present album name or rarely it can be empty/null on index 2 instead of track time
-    long duration;
-    try {
-      duration = DataFormatTools.durationTextToMillis(renderer.get("flexColumns")
-          .index(2)
-          .get("musicResponsiveListItemFlexColumnRenderer")
-          .get("text")
-          .get("runs")
-          .index(0)
-          .get("text").text());
-    } catch (RuntimeException ignored) {
-      duration = DataFormatTools.durationTextToMillis(renderer.get("flexColumns")
-          .index(3)
-          .get("musicResponsiveListItemFlexColumnRenderer")
-          .get("text")
-          .get("runs")
-          .index(0)
-          .get("text").text());
-    }
-    String videoId = renderer.get("doubleTapCommand")
-        .get("watchEndpoint")
-        .get("videoId").text();
 
     AudioTrackInfo info = new AudioTrackInfo(title, author, duration, videoId, false,
         WATCH_URL_PREFIX + videoId);
